LayoutTests/http/tests/webgpu/webgpu/compat/api/validation/createBindGroup-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: Importing binding name 'effectiveViewDimensionForTexture' is not found.

Harness Error (FAIL), message = Unhandled rejection: Importing binding name 'effectiveViewDimensionForTexture' is not found.
